+++ Plugins/iOS/FlurryConsent.h
@@ -1,18 +1,18 @@
 /*
- * Copyright 2018, Oath Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
+* Copyright 2018, Oath Inc.
+*
+* Licensed under the Apache License, Version 2.0 (the "License");
+* you may not use this file except in compliance with the License.
+* You may obtain a copy of the License at
+*
+*     http://www.apache.org/licenses/LICENSE-2.0
+*
+* Unless required by applicable law or agreed to in writing, software
+* distributed under the License is distributed on an "AS IS" BASIS,
+* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+* See the License for the specific language governing permissions and
+* limitations under the License.
+*/
 
 #import <Foundation/Foundation.h>
 
@@ -32,6 +32,7 @@
  *
  *  @param consentStrings          NSDictionary* <NSString*,NSString*> => <consent sting format , consent string>.
  *
+ *  @note  The values for ConsentStrings should be of correct base64url-encoded representation.
  *  @note  ConsentStrings must be provided if "isGDPRScope" is set to YES for the consent information to be valid
  *
  */